refactor: ExtractArbitraryPage
